// src/typings/balancePlatform/ibanAccountIdentification.ts
/*
 * The version of the OpenAPI document: v2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit this class manually.
 */


export class IbanAccountIdentification {
    /**
    * The international bank account number as defined in the [ISO-13616](https://www.iso.org/standard/81090.html) standard.
    */
    'iban': string;
    /**
    * **iban**
    */
    'type': IbanAccountIdentification.TypeEnum;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "iban",
            "baseName": "iban",
            "type": "string"
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "IbanAccountIdentification.TypeEnum"
        }    ];

    static getAttributeTypeMap() {
        return IbanAccountIdentification.attributeTypeMap;
    }
}

export namespace IbanAccountIdentification {
    export enum TypeEnum {
        Iban = 'iban'
    }
}